third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/declared/get-shorthand-expected.txt

This is a testharness.js-based test.
[FAIL] Getting a shorthand property that is partially set in css rule returns undefined
  assert_equals: Shorthand value must be undefined as it is not explicitly set expected (undefined) undefined but got (object) null
Harness: the test ran to completion.
